requirements.txt
ipyplot==1.0.5
numpy==1.23.2
Pillow==9.3.0
seaborn==0.12.2